+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/adp-tooling
 
+## 0.15.1
+
+### Patch Changes
+
+-   52ac1b0: feat: Add ADP manifest change wizard sub-generators
+
+## 0.15.0
+
+### Minor Changes
+
+-   e53c1f8: Remove archive.zip from deploy script to allow passing any of the additional CLI params.
+
 ## 0.14.44
 
 ### Patch Changes